fix: Add handling for non numeric inputs
In the performCalculation function, added a check to clear the calculator if only
negative signs have been entered (which would mean first and second input are equal to -).

This prevents the calculator returning NaN.

Also added a variable to store the operator, for the instance where a double negative returns a positive.

This affected the calculator display. i.e - -4 returned - +4 on the display.

This change now remedies that and returns the correct user input.